Switch to accountless implementation
Previously kept track of user accounts and assigned identities to
accounts. Although this could potentially enable some interesting
features, the complexity and corner cases were increasing, with
potential security implications as well.

Now there is no concept of users. There's a single "global" list
of identities, and if you can prove you control the provider
account for that identity, it can be tied to your login cookie
and used.