Fix running multiple instances on fly.io
Was having a problem where client app backends were sending
the token request to a different region than where the session
was originally created, causing it to fail. Currently using
fly.io's fly-replay functionality to get around that.

A generic solution in the future might be for instances to
transparently forward requests to other instances.