Make login cookie SameSite=None
This is currently necessary for FedCM. In order to try and protect
against unanticipated security issues from this, I added a new lax
cookie that is set on all requests. I then gated all non-FedCM access to
the login cookie behind the getLoginCookie function. This function
verifies that the lax cookie is present, which indicates the
request is not cross-site. We can't use the Origin header for this
because FedCM doesn't send it for privacy reasons.

Hopefully FedCM eventually changes to allow SameSite=lax so we
don't have to worry about this. See
https://github.com/w3c-fedid/FedCM/issues/587